requirements.txt
python-dotenv==1.0.0
pandas==2.0.2
datasets==2.13.1
huggingface-hub==0.15.1
